Weird temporary fix
Islandora API is wrong one derivative_alter hook but mostly checksum is
the faulty because it requires an abstract object to be passed as
parameter, even when not used (and we have no need for that here).
Since i will have to make a few pull requests agains tuque, islandora
and checksum to fix this, for now, to avoid errors with people having
checksum enabled, i created concrete class (don’t use it!) derived from
Abstract Object that implements the needed methods giving basically
null,s but avoiding to have a tuque connection, since i don’t want to
hit fedora at all.

All related to
https://github.com/mnylc/islandora_multi_importer/issues/11 so pinging
@patdunlavey  here.