For now i have to keep parent around
I thought about basically removing parent and parent generation based
on rows for updates, since that operation deals with PIDs that exist,
but that requires a rewrite on the preprocessor logic, where each row
needs a parent to be checked as valid. If not, it will be invalidated.
This at least allows to test and see how updating existing objects fail!