PHPExcel bug
PHPExcel can’t read values that have an html entity in the values… so
weird!!! So if you have a field that says like “hello<“ it dies bad bad
badly. This happens when your data comes from Google and IMI needs to
dump it to file. So maybe.. i can html entity it there, will need to
test and retest