Fake Object now can be in codependent Fake relation ships
But they are still fake, not healthy relationships can grown on fake
objects.

This allows Fake Object to at least survive a request for relationships
and simply return NULL or empty arrays for other use cases. #89 Would
need to revisited and probably a real object (such an overhead!) would
be necessary if our users use Context and it keeps adding more RDF
requirements.