Deal actively with MODS to DC
First attempt on triggering an internal MODS to DC in case the global
DC transformation is not enabled.
This is just blind code for now, have not tested this yet. It also
enables an empty collection policy to be attached to Collection CMODEL
objects. All this is only triggered on Ingests, not on Updates (for
obvious reasons), but we should, once tested, at least allow the DC
part to be triggered IF MODS is being updated.