Issue 95: i really like my DC to be generated please (#97)
* Possible fix for merge predicates on update

To avoid double membership issue

* debug

* debug bugged

* sameAs

* Now i got it right

* Deal with allowing same predicate many times

* Deal actively with MODS to DC

First attempt on triggering an internal MODS to DC in case the global
DC transformation is not enabled.
This is just blind code for now, have not tested this yet. It also
enables an empty collection policy to be attached to Collection CMODEL
objects. All this is only triggered on Ingests, not on Updates (for
obvious reasons), but we should, once tested, at least allow the DC
part to be triggered IF MODS is being updated.

* Ingest the created DC

* remove unused include

We are directly loading and applying XSLT without XML forms dependency
now. I mean, its just 4 lines

* Wrong logic!

addressing excellent review by @patdunlavey

* Another error. Please 2018 go away