Give a DS is source file name as label
This pull uses the file name as a DS label. There is some redundancy in the way i implemented this because i still want to get some feedback on the approach. Also, islandora_multi_importer_remote_file_get could eventually change the file name if someone is using a hook. Should we deal with that? Or assume that what the ingesting user intended the filename to be named is what goes into the DS Label?