Issue 62: Create a fake iterator for our fake Object (#105)
* Update FakeObject to match Iterator expectations

This should fix ISSUE #62
There is an issue with OHSP where the derivative triggering process tries (and does so) array push datastreams in place, which is wrong.
As stated there
https://github.com/Islandora/tuque/blob/ceaa130ab092dbad9d3c37073cba4575687d63ea/Object.php#L901-L903
Setting an offset is not the way, since you really want create a new datastream instead of pushing. But, anyway, we want this to play well. We will have to revision more times than expected this fakerObject class, since its only purpose is to allow Needed derivatives to be listed during the IMI setup and not to be used to add/remove Datastreams, there is already a CMODEL hook (see TECHMD) where you can add new Datstreams to an existing CMODEL without tinkering on existing objects.

* Coding Standard